+++ test/setup.js
@@ -1,4 +1,4 @@
-import { nodeAPI } from '../scripts/preload';
+import { nodeAPI } from '../main/preload';
 
 // inject the nodeAPI reference for testing
 global.window = {
